requirements.txt
django==3.1.14
djangorestframework==3.11.2
